Revert "Move the creation of aliases to async celery"
This reverts commit 54607ee2e7462abb246642c57c681456c0884e4c.
