+++ core/serializers.py
@@ -182,6 +182,7 @@ class OperationSerializer(serializers.ModelSerializer):
         operation.operation_query = query.build_query(operation=operation)
         operation.count_rows = True
         operation.save()
+        self.create_operation_data_aliases(operation)
 
         return operation
 
@@ -219,8 +220,43 @@ class OperationSerializer(serializers.ModelSerializer):
         instance.operation_query = query.build_query(operation=instance)
         instance.count_rows = True
         instance.save()
+        self.update_operation_data_aliases(instance)
 
         return instance
+
+    def create_operation_data_aliases(self, operation):
+        count, data = query.query_table(operation, 1, 0, estimate_count=True)
+        try:
+            data_column_keys = data[0].keys()
+            first_step = operation.get_operation_steps()[0]
+            columns = SourceColumnMap.objects.filter(source=first_step.source, name__in=data_column_keys)
+            for column in data_column_keys:
+                matching = columns.filter(name=column).first()
+                alias = self.create_operation_alias(operation, column, matching.alias if matching else column)
+                alias.save()
+        except: # FIXME: handle specific errors
+            pass
+
+    def update_operation_data_aliases(self, operation):
+        count, data = query.query_table(operation, 1, 0, estimate_count=True)
+        try:
+            data_column_keys = data[0].keys()
+            first_step = operation.get_operation_steps()[0]
+            columns = SourceColumnMap.objects.filter(source=first_step.source, name__in=data_column_keys)
+            # delete obsolete aliases
+            OperationDataColumnAlias.objects.filter(operation=operation).exclude(column_name__in=data_column_keys).delete()
+            for column in data_column_keys:
+                existing_alias = OperationDataColumnAlias.objects.filter(operation=operation, column_name=column).first()
+                if not existing_alias:
+                    matching = columns.filter(name=column).first()
+                    alias = self.create_operation_alias(operation, column, matching.alias if matching else column)
+                    alias.save()
+        except: # FIXME: handle specific errors
+            pass
+
+    def create_operation_alias(self, operation, column_name, column_alias):
+        return OperationDataColumnAlias.objects.create(
+            operation=operation, column_name=column_name, column_alias=column_alias)
 
 
 class ThemeSerializer(serializers.ModelSerializer):
