+++ frontend/models.py
@@ -3,14 +3,12 @@ from django.dispatch import receiver
 from django.core.signals import request_finished
 
 from core.models import Operation
-from core.tasks import count_operation_rows as count, create_operation_data_aliases as aliases
+from core.tasks import count_operation_rows as count
 
 
-@receiver(post_save, sender=Operation, dispatch_uid="operation__post_save")
-def operation_post_save(sender, instance=None, created=False, **kwargs):
-    if instance:
-        if instance.count_rows:
-            count.delay(instance.id)
-            aliases.delay(instance.id)
+@receiver(post_save, sender=Operation, dispatch_uid="count_operation_rows")
+def count_operation_rows(sender, instance=None, created=False, **kwargs):
+    if instance and instance.count_rows:
+        count.delay(instance.id)
 
-request_finished.connect(operation_post_save, dispatch_uid="operation__post_save")
+request_finished.connect(count_operation_rows, dispatch_uid="count_operation_rows")
